Extras/CDFDeprecated/Simulators/MP70Simulator/requirements.txt
boto3
requests
paho-mqtt
requests_aws_sign